ref: simplify navigation files
