+++ src/navigation/fishbot_navigation/__init__.py
@@ -1,4 +1,1 @@
 """Fishbot navigation package."""
-
-from .navigator import NavigatorNode  # noqa: F401
-
